fix: resolve QA review findings across translation correctness, security, and UX
Fixes 36 confirmed issues from a ruthless QA/security/performance review
(see QA_Review_Report.md), spanning:

Critical translation bugs:
- View/MV body regex converters no longer corrupt string literals or
  comments (mask/unmask string-literal-aware rewriting)
- WITH RECURSIVE and LIMIT/OFFSET now convert correctly for T-SQL targets
- Confidence scoring now reflects ERROR-severity warnings, not just the
  unsupported flag, so broken output no longer scores HIGH/PARTIAL
- Type-mapping YAML reverse-lookup collisions fixed (BigQuery INT64,
  Fabric DW SMALLINT/INT, Oracle NUMBER(p) precision)
- Inline column-level PRIMARY KEY now parsed correctly across all 9
  dialects (previously silently dropped)
- Redshift/Snowflake table names now correctly quoted in generated DDL
- ReDoS-class regex in Redshift DISTKEY/SORTKEY extraction replaced with
  a linear scan (~38s -> ~1s on a 3,200-column table); also fixed a
  pre-existing bug where table-level DISTKEY(col) could resolve to the
  wrong column entirely
- Added a request-size cap and moved transpilation to a thread pool so a
  large conversion can no longer block the whole server

Additional fixes: DECIMAL precision clamping, DISTKEY casing, BigQuery
FK/UNIQUE/CHECK constraints, Databricks PK "NOT ENFORCED" + IDENTITY
type widening, Oracle DATE time-component warning, unified the two
divergent SQL statement splitters, object_type request validation,
broken start-*.cmd scripts, de-hardcoded personal paths from docs,
added LICENSE/CONTRIBUTING.md/CI workflow, pinned backend dependencies,
removed orphaned schema-inference dead code, and accessibility fixes
(aria-expanded, aria-modal, focus management, missing dialect logo).

Full backend test suite (3,986 tests) passes; golden snapshots
regenerated where output intentionally changed to fix a bug (each
diff verified against the specific bug it corrects).

+++ backend/tests/golden/bigquery_to_databricks/procedure.sql
@@ -1,7 +1,7 @@
 -- Databricks does NOT support CREATE PROCEDURE.
 -- Docs: https://docs.databricks.com/en/sql/language-manual/
 -- Converted to a SQL UDF. Body requires significant manual adaptation.
-CREATE OR REPLACE FUNCTION `analytics`.`upsert_order`(p_order_id TINYINT, p_amount DECIMAL)
+CREATE OR REPLACE FUNCTION `analytics`.`upsert_order`(p_order_id BIGINT, p_amount DECIMAL)
 RETURNS STRING
 RETURN (
   -- -- Translated from unknown → databricks
