+++ backend/tests/golden/bigquery_to_databricks/table.sql
@@ -1,6 +1,6 @@
 CREATE OR REPLACE TABLE `analytics`.`orders` (
-    `order_id` TINYINT NOT NULL,
-    `customer_id` TINYINT NOT NULL,
+    `order_id` BIGINT NOT NULL,
+    `customer_id` BIGINT NOT NULL,
     `amount` DECIMAL(18,2) NOT NULL,
     `status` CHAR DEFAULT 'pending',
     `created_at` TIMESTAMP NOT NULL
