+++ backend/tests/golden/bigquery_to_fabric_dw/table.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS [analytics].[orders];
 CREATE TABLE [analytics].[orders] (
-    [order_id] SMALLINT NOT NULL,
-    [customer_id] SMALLINT NOT NULL,
+    [order_id] BIGINT NOT NULL,
+    [customer_id] BIGINT NOT NULL,
     [amount] DECIMAL(18,2) NOT NULL,
     [status] CHAR,
     [created_at] DATETIME2(6) NOT NULL
