+++ backend/tests/golden/bigquery_to_sqlserver/table.sql
@@ -1,8 +1,8 @@
 DROP TABLE IF EXISTS [analytics].[orders];
 GO
 CREATE TABLE [analytics].[orders] (
-    [order_id] TINYINT NOT NULL,
-    [customer_id] TINYINT NOT NULL,
+    [order_id] BIGINT NOT NULL,
+    [customer_id] BIGINT NOT NULL,
     [amount] DECIMAL(18,2) NOT NULL,
     [status] CHAR DEFAULT 'pending',
     [created_at] DATETIMEOFFSET NOT NULL
